Added clean output if info plus additional options
I removed the `var_dump($data)` and added in actual output of the widget, depending on what options are set by the user. I also added in additional options for the user to pick how many plugins to show, as well as checkboxes to show or hide the plugin info (version, date added, last updated, downloads).